Updated schema jar and moved e2e to explicit package
