+++ src/endToEndTest/java/org/radarcns/endtoend/util/Sender.java
@@ -1,4 +1,4 @@
-package org.radarcns.util;
+package org.radarcns.endtoend.util;
 
 import java.io.IOException;
 import java.util.Arrays;
